+++ test/libsolidity/SolidityNatspecJSON.cpp
@@ -378,7 +378,7 @@ BOOST_AUTO_TEST_CASE(dev_return)
 	"            \"a\": \"Documentation for the first parameter starts here. Since it's a really complicated parameter we need 2 lines\",\n"
 	"            \"second\": \"Documentation for the second parameter\"\n"
 	"        },\n"
-	"        \"return\": {\n"
+	"        \"returns\": {\n"
 	"            \"d\": \"The result of the multiplication\"\n"
 	"        }\n"
 	"    }\n"
@@ -411,8 +411,81 @@ BOOST_AUTO_TEST_CASE(dev_return_desc_after_nl)
 	"            \"a\": \"Documentation for the first parameter starts here. Since it's a really complicated parameter we need 2 lines\",\n"
 	"            \"second\": \"Documentation for the second parameter\"\n"
 	"        },\n"
-	"        \"return\": {\n"
+	"        \"returns\": {\n"
 	"            \"d\": \"The result of the multiplication\"\n"
+	"        }\n"
+	"    }\n"
+	"}}";
+
+	checkNatspec(sourceCode, "test", natspec, false);
+}
+
+BOOST_AUTO_TEST_CASE(dev_return_desc_multiple_unamed_mixed)
+{
+	char const* sourceCode = R"(
+		contract test {
+			/// @dev Multiplies a number by 7 and adds second parameter
+			/// @param a Documentation for the first parameter starts here.
+			/// Since it's a really complicated parameter we need 2 lines
+			/// @param second Documentation for the second parameter
+			/// @return The result of the multiplication
+			/// @return _cookies And cookies with nutella
+			function mul(uint a, uint second) public returns (uint, uint _cookies) {
+				uint mul = a * 7;
+				return (mul, second);
+			}
+		}
+	)";
+
+	char const* natspec = "{"
+	"\"methods\":{"
+	"    \"mul(uint256,uint256)\":{ \n"
+	"        \"details\": \"Multiplies a number by 7 and adds second parameter\",\n"
+	"        \"params\": {\n"
+	"            \"a\": \"Documentation for the first parameter starts here. Since it's a really complicated parameter we need 2 lines\",\n"
+	"            \"second\": \"Documentation for the second parameter\"\n"
+	"        },\n"
+	"        \"returns\": {\n"
+	"            \"_0\": \"The result of the multiplication\",\n"
+	"            \"_cookies\": \"And cookies with nutella\"\n"
+	"        }\n"
+	"    }\n"
+	"}}";
+
+	checkNatspec(sourceCode, "test", natspec, false);
+}
+
+BOOST_AUTO_TEST_CASE(dev_return_desc_multiple_unamed_mixed_2)
+{
+	char const* sourceCode = R"(
+		contract test {
+			/// @dev Multiplies a number by 7 and adds second parameter
+			/// @param a Documentation for the first parameter starts here.
+			/// Since it's a really complicated parameter we need 2 lines
+			/// @param second Documentation for the second parameter
+			/// @return _cookies And cookies with nutella
+			/// @return The result of the multiplication
+			/// @return _milk And milk with nutella
+			function mul(uint a, uint second) public returns (uint _cookies, uint, uint _milk) {
+				uint mul = a * 7;
+        uint milk = 4;
+				return (mul, second, milk);
+			}
+		}
+	)";
+
+	char const* natspec = "{"
+	"\"methods\":{"
+	"    \"mul(uint256,uint256)\":{ \n"
+	"        \"details\": \"Multiplies a number by 7 and adds second parameter\",\n"
+	"        \"params\": {\n"
+	"            \"a\": \"Documentation for the first parameter starts here. Since it's a really complicated parameter we need 2 lines\",\n"
+	"            \"second\": \"Documentation for the second parameter\"\n"
+	"        },\n"
+	"        \"returns\": {\n"
+	"            \"_cookies\": \"And cookies with nutella\",\n"
+	"            \"_1\": \"The result of the multiplication\",\n"
+	"            \"_milk\": \"And milk with nutella\"\n"
 	"        }\n"
 	"    }\n"
 	"}}";
@@ -445,9 +518,9 @@ BOOST_AUTO_TEST_CASE(dev_return_desc_multiple_unamed)
 	"            \"a\": \"Documentation for the first parameter starts here. Since it's a really complicated parameter we need 2 lines\",\n"
 	"            \"second\": \"Documentation for the second parameter\"\n"
 	"        },\n"
-	"        \"return\": {\n"
-	"            \"_1\": \"The result of the multiplication\",\n"
-	"            \"_2\": \"And cookies with nutella\"\n"
+	"        \"returns\": {\n"
+	"            \"_0\": \"The result of the multiplication\",\n"
+	"            \"_1\": \"And cookies with nutella\"\n"
 	"        }\n"
 	"    }\n"
 	"}}";
@@ -466,7 +539,7 @@ BOOST_AUTO_TEST_CASE(dev_return_desc_multiple)
 			/// @return d The result of the multiplication
 			/// @return f And cookies with nutella
 			function mul(uint a, uint second) public returns (uint d, uint f) {
-        uint mul = a * 7;
+				uint mul = a * 7;
 				return (mul, second);
 			}
 		}
@@ -480,7 +553,7 @@ BOOST_AUTO_TEST_CASE(dev_return_desc_multiple)
 	"            \"a\": \"Documentation for the first parameter starts here. Since it's a really complicated parameter we need 2 lines\",\n"
 	"            \"second\": \"Documentation for the second parameter\"\n"
 	"        },\n"
-	"        \"return\": {\n"
+	"        \"returns\": {\n"
 	"            \"d\": \"The result of the multiplication\",\n"
 	"            \"f\": \"And cookies with nutella\"\n"
 	"        }\n"
@@ -514,7 +587,7 @@ BOOST_AUTO_TEST_CASE(dev_multiline_return)
 	"            \"a\": \"Documentation for the first parameter starts here. Since it's a really complicated parameter we need 2 lines\",\n"
 	"            \"second\": \"Documentation for the second parameter\"\n"
 	"        },\n"
-	"        \"return\": {\n"
+	"        \"returns\": {\n"
 	"            \"d\": \"The result of the multiplication and cookies with nutella\",\n"
 	"        }\n"
 	"    }\n"
@@ -549,13 +622,28 @@ BOOST_AUTO_TEST_CASE(dev_multiline_comment)
 	"            \"a\": \"Documentation for the first parameter starts here. Since it's a really complicated parameter we need 2 lines\",\n"
 	"            \"second\": \"Documentation for the second parameter\"\n"
 	"        },\n"
-	"        \"return\": {\n"
+	"        \"returns\": {\n"
 	"            \"d\": \"The result of the multiplication and cookies with nutella\",\n"
 	"        }\n"
 	"    }\n"
 	"}}";
 
 	checkNatspec(sourceCode, "test", natspec, false);
+}
+
+BOOST_AUTO_TEST_CASE(dev_documenting_no_return_paramname)
+{
+	char const* sourceCode = R"(
+		contract test {
+			/// @dev Multiplies a number by 7 and adds second parameter
+			/// @param a Documentation for the first parameter
+			/// @param second Documentation for the second parameter
+			/// @return
+			function mul(uint a, uint second) public returns (uint d) { return a * 7 + second; }
+		}
+	)";
+
+	expectNatspecError(sourceCode);
 }
 
 BOOST_AUTO_TEST_CASE(dev_contract_no_doc)
@@ -871,7 +959,7 @@ BOOST_AUTO_TEST_CASE(dev_constructor_and_function)
 				"a" : "Documentation for the first parameter starts here. Since it's a really complicated parameter we need 2 lines",
 				"second" : "Documentation for the second parameter"
 			},
-			"return" : {
+			"returns" : {
 				"d": "The result of the multiplication and cookies with nutella"
 			}
 		},